+++ scanpipe/tests/regen_test_data.py
@@ -48,11 +48,18 @@ class RegenTestData(TestCase):
     $ ./manage.py test --pattern "regen*.py"
 
     - Docker:
-    $ docker compose run --volume "$(pwd)/scanpipe/":/opt/scancodeio/scanpipe/ web \
-        ./manage.py test --pattern "regen*.py"
+    $ docker compose run --rm --volume "$(pwd)/scanpipe/":/opt/scancodeio/scanpipe/ \
+        web ./manage.py test --pattern "regen*.py"
 
     Warning: Once the test data is updated, run the whole test suite with the
     `SCANCODEIO_TEST_FIXTURES_REGEN` setting enabled to regen the expected files.
+
+    - Local:
+    $ SCANCODEIO_TEST_FIXTURES_REGEN=True ./manage.py test
+
+    - Docker:
+    $ docker compose run --rm --volume "$(pwd)/scanpipe/":/opt/scancodeio/scanpipe/ \
+        -e SCANCODEIO_TEST_FIXTURES_REGEN=True web ./manage.py test
     """
 
     data_location = Path(__file__).parent / "data"
@@ -73,10 +80,16 @@ class RegenTestData(TestCase):
         self.assertEqual(0, exitcode)
 
         # ScanCode-toolkit scan result
-        scan_options = ["--copyright", "--info", "--license", "--package"]
-        scan_location = str(project1.codebase_path)
-        output_location = str(self.data_location / "asgiref-3.3.0_toolkit_scan.json")
-        scancode.run_scancode(scan_location, output_location, scan_options)
+        scancode.run_scan(
+            location=str(project1.codebase_path),
+            output_file=str(self.data_location / "asgiref-3.3.0_toolkit_scan.json"),
+            run_scan_args={
+                "copyright": True,
+                "info": True,
+                "license": True,
+                "package": True,
+            },
+        )
 
         # ScanCode.io results
         test_file_location = self.data_location / "asgiref-3.3.0_scanpipe_output.json"
@@ -124,10 +137,15 @@ class RegenTestData(TestCase):
         project2.copy_input_from(input_location)
         input.copy_input(input_file, project2.codebase_path)
         scancode.extract_archives(location=project2.codebase_path)
-        scan_options = ["--info", "--package"]
         scan_location = str(
             project2.codebase_path / "package_assembly_codebase.tar.gz-extract/"
         )
         json_filename = "package_assembly_codebase.json"
-        output_location = str(self.data_location / "scancode" / json_filename)
-        scancode.run_scancode(scan_location, output_location, scan_options)
+        scancode.run_scan(
+            location=scan_location,
+            output_file=str(self.data_location / "scancode" / json_filename),
+            run_scan_args={
+                "info": True,
+                "package": True,
+            },
+        )