+++ CHANGELOG.rst
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+v33.0.0 (unreleased)
+--------------------
+
+- Upgrade Django to version 5.0 and drop support for Python 3.8 and 3.9
+  https://github.com/nexB/scancode.io/issues/1020
+
+- Refactor run_scancode to not fail on scan errors happening at the resource level,
+  such as a timeout. Project error message are created instead.
+  https://github.com/nexB/scancode.io/issues/1018
+
+- Add support for the SCANCODEIO_SCAN_FILE_TIMEOUT setting in the scan_package pipeline.
+  https://github.com/nexB/scancode.io/issues/1018
+
+- Add support for non-archive single file in the scan_package pipeline.
+  https://github.com/nexB/scancode.io/issues/1009
+
 v32.7.0 (2023-10-25)
 --------------------